version quasiment finie avec PID en comment

+++ Projet_Movo_bot/Movo_bot/accelerometre.h
@@ -4,7 +4,7 @@
 #include <sensors/imu.h>
 
 void start_accelerometre(void);
-float get_acceleration_y(void);
+int get_acceleration_y(void);
 void get_gravity(imu_msg_t *imu_values);
 void wait_accelerometre_mesure(void);
 uint8_t get_impact(void);
